refactor: layout portfolio dashboard
